% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_time_n.R
\name{extreme_time_n}
\alias{extreme_time_n}
\title{Compute Outlier Statistics and Time-based Metrics}
\usage{
extreme_time_n(data, days, ref_date = Sys.Date())
}
\arguments{
\item{data}{data.table, required.
The input data table which must contain the following columns: \code{visit_time} (POSIXct), \code{location} (character),
\code{responder} (character), \code{weight} (numeric, in grams), and \code{duration} (numeric, in seconds).}

\item{days}{integer, required.
The number of days before the reference date to include in the analysis.}

\item{ref_date}{Date, optional.
The specific date to use as the reference for filtering data; defaults to the current system date.}
}
\value{
A list of two data.tables:
\itemize{
\item \code{weight_outlier}: A data table in wide format showing the percentage of weight outliers
per location across the specified dates.
\item \code{feedintake}: A data table in wide format showing the count of observations and total
duration in hours per location across the specified dates.
}
}
\description{
This function analyzes weight data to identify outliers using the Interquartile Range (IQR) method,
and computes time-based metrics for data entries within a specified number of days from a reference date.
It also reshapes the results into a wide format, separating the percentage of weight outliers and other metrics
like count of observations and total duration by location and date.
}
\note{
The function assumes that the input data has all the required columns. The 'visit_time' should
be in POSIXct format to correctly convert to dates. Make sure the data is pre-cleaned to avoid errors
due to missing values, especially in the 'location' and 'responder' columns.
}
\examples{
# Load CSV data
data <- data.table::fread("C:/Users/Dell/Documents/projects/pptsdm_data/ppt_monitor_test_data.csv")
print(extreme_time_n(data = data, days = 5))
}
