% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fid_monitor.R
\name{fid_monitor}
\alias{fid_monitor}
\title{Feed intake monitor of pig performance test station}
\usage{
fid_monitor(data, station_type, save_path)
}
\arguments{
\item{data}{Csv data of pig performance test station}

\item{station_type}{Type of pig performance test station, you can choose nedap or fire}

\item{save_path}{A string specifying the output file path}
}
\value{
A pdf file about feed intake data monitor
}
\description{
Feed intake monitor of pig performance test station
}
\examples{
data <- data.table::fread("C:/Users/Dell/Downloads/test/ppt_test_data.csv")
fid_monitor(data = data, station_type = "nedap", save_path = "C:/Users/Dell/Downloads/test")
}
