% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_monitor.R
\name{growth_monitor}
\alias{growth_monitor}
\title{Feed intake monitor of pig performance test station}
\usage{
growth_monitor(
  data,
  lmrob = T,
  threshold = 0.8,
  my_break = NULL,
  range_offset = 0.5,
  save_path
)
}
\arguments{
\item{data}{Csv data of pig performance test station}

\item{lmrob}{Default lmrob = TRUE}

\item{threshold}{Defult threshold is 0.8}

\item{my_break}{Default break is NULL}

\item{range_offset}{Default range_offset is 0.5}

\item{save_path}{A string specifying the output file path}
}
\value{
png files about growth curve monitor in each station
}
\description{
Feed intake monitor of pig performance test station
}
\examples{
data <- data.table::fread("C:/Users/Dell/Downloads/test/ppt_test_data.csv")
growth_monitor(data = data, save_path = "C:/Users/Dell/Downloads/test")
}
