% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_weight.R
\name{house_weight}
\alias{house_weight}
\title{Calculate Mean Weight for Specified Location and Date Range}
\usage{
house_weight(data, house_width = "1", days, ref_date = Sys.Date())
}
\arguments{
\item{data}{A data frame that must include the columns: "location", "date", and "weight".}

\item{house_width}{A string indicating the number of characters to use from the 'location'
column to create the 'house' identifier. Defaults to "1".}

\item{days}{An integer specifying the number of days to include in the analysis
back from the reference date.}

\item{ref_date}{A Date object used as the reference date for filtering data.
Defaults to the current system date.}
}
\value{
Returns a data frame in wide format where each column represents a house
and each row represents a date within the specified range. Values are mean weights
(in kilograms, rounded to two decimal places) for each house-date combination.
}
\description{
This function processes a dataset to extract weights by specified house (part of location),
calculates mean weights per house over a given number of days from a reference date, and
returns the data in a wide format with dates as rows and houses as columns.
}
\note{
If 'data' does not contain the necessary columns, the function will stop with an error message.
It is crucial to ensure that the 'location' and 'date' columns are formatted correctly.
}
\examples{
# Load CSV data
data <- data.table::fread("C:/Users/Dell/Documents/projects/pptsdm_data/ppt_monitor_test_data.csv")
head(house_weight(data = data, days = 5))
}
