% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_feedintake.R
\name{low_feedintake}
\alias{low_feedintake}
\title{Analyze Low Feed Intake}
\usage{
low_feedintake(data, days, ref_date = Sys.Date())
}
\arguments{
\item{data}{A data table that must include the columns: 'location', 'date', 'responder',
and 'feed_intake'. The 'visit_time' must be able to be converted to IDateTime format.}

\item{days}{An integer specifying the number of days to include in the analysis up to and including
the reference date.}

\item{ref_date}{A Date object used as the reference date for filtering data; defaults to the current
system date.}
}
\value{
Returns a data table in wide format where each row represents a combination of location and
responder, columns are dates, and values are the summed feed intake that are below the threshold
of 0.5 kg. Also includes a 'sum_feedintake' column which is the total sum of feed intake across
the period for each row.
}
\description{
This function processes a dataset to calculate the total feed intake per location, date,
and responder. It then identifies days where the sum of feed intake is below a threshold,
specifically less than 0.5 kg. The results are returned in a wide format with each location
and responder combination as a row and dates as columns.
}
\note{
This function assumes that the input data has been preprocessed to include the necessary columns.
It will stop and raise an error if the 'visit_time' cannot be converted to date and time format.
}
\examples{
# Load CSV data
data <- data.table::fread("C:/Users/Dell/Documents/projects/pptsdm_data/ppt_monitor_test_data.csv")
print(low_feedintake(data = data, days = 5))
}
