% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_all.R
\name{monitor_all}
\alias{monitor_all}
\title{Monitor all by hands}
\usage{
monitor_all(data, begin_date, house_width = "1", days, save_path)
}
\arguments{
\item{data}{data.frame, required. The input data frame containing the data to be processed.}

\item{begin_date}{An optional Date object or character string specifying the beginning date for the data to be processed. If not provided, all dates in the data will be considered.}

\item{house_width}{character, optional. A character string representing the house width. Default is "1".}

\item{days}{integer, optional. An integer representing the number of days for the analysis.}

\item{save_path}{character, required. A character string representing the path where the output file will be saved.}
}
\description{
Combine the \code{fid_monitor}, \code{station_monitor}, and \code{table_monitor} functions together.
}
\examples{
# Load CSV data
data <- data.table::fread("C:/Users/Dell/Documents/projects/pptsdm_data/ppt_monitor_test_data.csv")
monitor_all(data, begin_date = "2024-05-01", days = 5, save_path = "C:/Users/Dell/Downloads/test")
}
