% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_schedule.R
\name{monitor_schedule}
\alias{monitor_schedule}
\title{Feed intake monitor of pig performance test station}
\usage{
monitor_schedule(
  taskname,
  schedule,
  starttime,
  startdate,
  rscript_args = NULL,
  ...
)
}
\arguments{
\item{taskname}{A character string with the name of the task. Defaults to the filename. Should not contain any spaces}

\item{schedule}{Either one of 'ONCE', 'MONTHLY', 'WEEKLY', 'DAILY', 'HOURLY', 'MINUTE', 'ONLOGON', 'ONIDLE}

\item{starttime}{A timepoint in HH:mm format indicating when to run the script. Defaults to within 62 seconds}

\item{startdate}{A date that specifies the first date on which to run the task. Only applicable if schedule is of type 'MONTHLY', 'WEEKLY', 'DAILY', 'HOURLY', 'MINUTE'. Defaults to today in '\%d/\%m/\%Y' format. Change to your locale format if needed}

\item{rscript_args}{Character string with further arguments passed on to Rscript}

\item{...}{other parameters}
}
\value{
pdf and pngs
}
\description{
Feed intake monitor of pig performance test station
}
\examples{
monitor_schedule(
  taskname = "ppt_csv_monitor",
  schedule = "DAILY",
  starttime = "10:05",
  startdate = format(Sys.Date(), "\%Y/\%m/\%d"),
  rscript_args = list(csv_path = "C:/Users/Dell/Downloads/test",
                      save_path = "C:/Users/Dell/Downloads/test"))
}
