% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_location.R
\name{number_location}
\alias{number_location}
\title{Count Responders by Location and Date}
\usage{
number_location(data, house_width, days, ref_date = Sys.Date())
}
\arguments{
\item{data}{A data table that must include the columns: 'location', 'date', 'responder'.
The function expects that 'date' can be converted into Date format.}

\item{house_width}{An integer specifying the number of characters from the start of the 'location'
field to define a 'house'.}

\item{days}{An integer specifying the number of days to include in the analysis back from the
reference date.}

\item{ref_date}{A Date object used as the reference date for filtering data; defaults to the current
system date.}
}
\value{
Returns a data table in wide format where each row represents a date and each column a house.
Each entry is the count of responders for that house and date.
}
\description{
This function processes a dataset to calculate the number of responders per specified 'house'
(a subset of the 'location' field) for each date within a specified number of days from a reference date.
The data is first processed to ensure the 'date' column is in Date format, 'house' is extracted from
'location', and duplicates are removed. The output is provided in wide format with dates as rows
and houses as columns.
}
\note{
This function depends on another function 'process data' to first ensure the data is in the correct
format. It is crucial that the input data and the 'process_data' function are properly configured
to handle the data correctly.
}
\examples{
# Load CSV data
data <- data.table::fread("C:/Users/Dell/Documents/projects/pptsdm_data/ppt_monitor_test_data.csv")
print(number_location(data = data, house_width = "1", days = 5))
}
