% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process Data Frame or Data Table}
\usage{
process_data(data)
}
\arguments{
\item{data}{A data frame or data table that must contain specific columns with specific data types.
Required columns are: "animal_number", "lifenumber", "responder", "location", "visit_time",
"duration", "state", "weight", "feed_intake".}
}
\value{
Returns a modified data table with computed sequence features and potentially other transformations.
The function ensures that columns are of the correct type and that data integrity is maintained.
}
\description{
This function performs various checks and transformations on a provided data frame or data table.
It checks for necessary columns, correct data types, and performs various operations like filtering,
joining, and calculating sequences.
}
\note{
This function will stop execution and produce an error if the input data does not meet the expected
requirements, such as missing columns or incorrect data types.
}
\examples{
# Load CSV data
data <- data.table::fread("C:/Users/Dell/Documents/projects/pptsdm_data/ppt_monitor_test_data.csv")
processed_data <- process_data(data = data)
head(processed_data)
}
