% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_monitor.R
\name{table_monitor}
\alias{table_monitor}
\title{Table feed intake monitor of pig performance test station}
\usage{
table_monitor(data, house_width = "1", days = 7, save_path)
}
\arguments{
\item{data}{A data frame or data table. This is the data to be processed. It should have specific columns depending on the station type.}

\item{house_width}{A character or numeric value specifying the width of the house number.If numeric, it will be converted to a character string. default house_width is 1.}

\item{days}{Days you can choose to monitor latest n days data, default days is 7.}

\item{save_path}{A character string specifying the path where the output PNG files will be saved.}
}
\value{
A list of data.tables, each representing different aspects of station monitoring results. This list is also saved to an Excel file at the specified \code{save_path}.
}
\description{
Table feed intake monitor of pig performance test station
}
\examples{
# Load CSV data
data <- data.table::fread("C:/Users/Dell/Desktop/test/monitor_test_data.csv")
# Monitor station and data
res <- table_monitor(data = data, save_path = "C:/Users/Dell/Desktop/test")
# Monitor the number of times 'na' appears in the last 7 days
res$responder_na
# Monitor the percentage of extreme weight records in the last 7 days
res$extreme_weight
# Monitor the visiting time and frequency of pigs in the last 7 days
res$feed_time_n
# Monitor the total feed intake over the last 7 days
res$all_feedintake
# Monitor the average feed intake over the last 7 days
res$mean_feedintake
# Monitor the average weight per pen over the last 7 days
res$house_weight
# Monitor visit time in each hour over the last 1 day.
res$visit_n
# Monitor feed intake time in each hour over the last 1 day.
res$feed_time
# Monitor feed intake in each hour over the last 1 day.
res$feed_intake
}
