% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fid_monitor.R
\name{fid_monitor}
\alias{fid_monitor}
\title{Feed intake monitor of pig performance test station}
\usage{
fid_monitor(data, begin_date = NULL, station_type, save_path)
}
\arguments{
\item{data}{A data frame or data table containing the nedap or fire pig performance test data to be processed. Columns must include 'visit_time', 'location', 'responder', 'feed_intake'.}

\item{begin_date}{An optional Date object or character string specifying the beginning date for the data to be processed. If not provided, all dates in the data will be considered.}

\item{station_type}{A character string specifying the type of station. This must be either 'nedap' or 'fire'.}

\item{save_path}{A character string specifying the path where the output PDF will be saved.}
}
\value{
This function does not return a value. It saves a PDF file to the specified path.
}
\description{
Feed intake monitor of pig performance test station
}
\examples{
# Load CSV data
data <- data.table::fread("C:/Users/Dell/Documents/projects/pptsdm_data/ppt_monitor_test_data.csv")
# Feed intake monitor
fid_monitor(data = data, station_type = "nedap", save_path = "C:/Users/Dell/Downloads/test")
}
