% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_monitor.R
\name{station_monitor}
\alias{station_monitor}
\title{Feed intake monitor of pig performance test station}
\usage{
station_monitor(data, begin_date = NULL, station_type, save_path)
}
\arguments{
\item{data}{A data frame or data table. This is the data to be processed. It should have specific columns depending on the station type.}

\item{begin_date}{An optional parameter. If provided, only data from this date onwards will be considered. It can be a Date object or a character string in the form 'yyyy-mm-dd'. Default is NULL, which means all dates in the data will be considered.}

\item{station_type}{A character string specifying the type of station. This must be either 'nedap' or 'fire'.}

\item{save_path}{A character string specifying the path where the output PNG files will be saved.}
}
\value{
This function does not return a value. It saves PNG files to the specified path.
}
\description{
Feed intake monitor of pig performance test station
}
\examples{
# Load CSV data
data <- data.table::fread("C:/Users/Dell/Documents/projects/pptsdm_data/ppt_monitor_test_data.csv")
# Station monitor
station_monitor(data = data, station_type = "nedap", save_path = "C:/Users/Dell/Downloads/test")
}
